NetworkPopulations

Documentation for NetworkPopulations.

NetworkPopulations.InitMode

NetworkPopulations.McmcInitialiser

NetworkPopulations.ScaledBeta

NetworkPopulations.TrGeometric

NetworkPopulations.KL_error

NetworkPopulations.KL_error

NetworkPopulations.KL_error

NetworkPopulations.cardinality

NetworkPopulations.cardinality

NetworkPopulations.draw_sample

NetworkPopulations.draw_sample

NetworkPopulations.draw_sample!

NetworkPopulations.draw_sample!

NetworkPopulations.draw_sample!

NetworkPopulations.draw_sample!

NetworkPopulations.draw_sample!

NetworkPopulations.eachinterseq

NetworkPopulations.get_normalising_const

NetworkPopulations.get_normalising_const

NetworkPopulations.get_path_sequences

NetworkPopulations.get_sample_space

NetworkPopulations.get_sample_space

NetworkPopulations.get_vertex_counts

NetworkPopulations.multigraph_edge_list

NetworkPopulations.pmf_unormalised

NetworkPopulations.string_to_int

NetworkPopulations.InitMode — Type

InitMode <: McmcInitialiser - this is a MCMC initialisation scheme for SIS model samplers which starts the MCMC chain at the model mode by default.

NetworkPopulations.McmcInitialiser — Type

Abstract type representing initialisation schemes for SIS model samplers.

NetworkPopulations.ScaledBeta — Type

Scaled Beta distribution. Instantiate with ScaledBeta() e.g.

d = ScaledBeta(α, β, 0.0, 10.0)

would make a ScaledBeta type d which is bounded by 0.0 and 10.0. Note

(α,β) are usual parameters;

lb = lower bound;

ub = upper bound.

NetworkPopulations.TrGeometric — Type

TrGeomtric(p::Float64, lb::Int, ub::Int)

Truncated Geometric distribution. Note:

lb = lower bound = samllest possible value (default=0)

ub = upper bound = largest possible value.

Instantiation:

TrGeometric(p, ub) - here lb defaults to 0

TrGeometric(p, lb, ub) here lb is specified

NetworkPopulations.KL_error — Method

KL_error(output::SimMcmcOutput, true_dist::Dict)

Takes the true distribution, as represented by a dictionary, as an additional parameter.

NetworkPopulations.KL_error — Method

KL_error(output::SisMcmcOutput, true_dist::Dict)

Takes the true distribution, as represented by a dictionary, as an additional parameter.

NetworkPopulations.KL_error — Method

KL_error(output::SisMcmcOutput)

Find the KL-divergence between the true distribution and the approximation obtained via the MCMC sample output.

That is, if $π(\mathcal{S})$ is the true distribution and $\tilde{\pi}(\mathcal{S})$ is the approximation then this function will return

\[\sum_{\mathcal{S}} \tilde{\pi}(\mathcal{S}) \log\left( \frac{\tilde{\pi}(\mathcal{S})}{\pi(\mathcal{S})}\right)\]

where we assume that $0 \cdot \log(0) = 0$.

NetworkPopulations.cardinality — Method

Calculate the sample space cardinality.

NetworkPopulations.cardinality — Method

Calculate the sample space cardinality.

NetworkPopulations.draw_sample! — Method

draw_sample!(
    sample_out::InteractionSequenceSample, 
    mcmc::SimMcmcInsertDeleteProportional, 
    model::SIM;
    burn_in::Int=mcmc.burn_in,
    lag::Int=mcmc.lag,
    init::InteractionSequence=get_init(model, mcmc.init)
)

Draw sample in-place from given SIM model model::SIM via MCMC algorithm with edit allocation and interaction insertion/deletion, storing output in sample_out::InteractionSequenceSample.

Accepts keyword arguments to change MCMC output, including burn-in, lag and initial values. If not given, these are set to the default values of the passed MCMC sampler mcmc::SimMcmcInsertDelete.

NetworkPopulations.draw_sample! — Method

draw_sample!(
    sample_out::InteractionSequenceSample, 
    mcmc::SimMcmcInsertDeleteSubpath, 
    model::SIM;
    burn_in::Int=mcmc.burn_in,
    lag::Int=mcmc.lag,
    init::InteractionSequence=get_init(model, mcmc.init)
)

Draw sample in-place from given SIM model model::SIM via MCMC algorithm with edit allocation and interaction insertion/deletion, storing output in sample_out::InteractionSequenceSample.

Accepts keyword arguments to change MCMC output, including burn-in, lag and initial values. If not given, these are set to the default values of the passed MCMC sampler mcmc::SimMcmcInsertDelete.

NetworkPopulations.draw_sample! — Method

draw_sample!(
    sample_out::InteractionSequenceSample, 
    mcmc::SisMcmcInsertDelete, 
    model::SIS;
    burn_in::Int=mcmc.burn_in,
    lag::Int=mcmc.lag,
    init::InteractionSequence=get_init(model, mcmc.init)
)

Draw sample in-place from given SIS model model::SIS via MCMC algorithm with edit allocation and interaction insertion/deletion, storing output in sample_out::InteractionSequenceSample.

Accepts keyword arguments to change MCMC output, including burn-in, lag and initial values. If not given, these are set to the default values of the passed MCMC sampler mcmc::SisMcmcInsertDelete.

NetworkPopulations.draw_sample! — Method

draw_sample!(
    sample_out::InteractionSequenceSample, 
    mcmc::SisMcmcSplitMerge, 
    model::SIS;
    burn_in::Int=mcmc.burn_in,
    lag::Int=mcmc.lag,
    init::InteractionSequence=get_init(model, mcmc.init)
)

Draw sample in-place from given SIS model model::SIS via MCMC algorithm with edit allocation and interaction insertion/deletion, storing output in sample_out::InteractionSequenceSample.

Accepts keyword arguments to change MCMC output, including burn-in, lag and initial values. If not given, these are set to the default values of the passed MCMC sampler mcmc::SisMcmcInsertDelete.

NetworkPopulations.draw_sample! — Method

draw_sample!(
    sample_out::InteractionSequenceSample, 
    mcmc::SimMcmcInsertDelete, 
    model::SIM;
    burn_in::Int=mcmc.burn_in,
    lag::Int=mcmc.lag,
    init::InteractionSequence=get_init(model, mcmc.init)
)

Draw sample in-place from given SIM model model::SIM via MCMC algorithm with edit allocation and interaction insertion/deletion, storing output in sample_out::InteractionSequenceSample.

Accepts keyword arguments to change MCMC output, including burn-in, lag and initial values. If not given, these are set to the default values of the passed MCMC sampler mcmc::SimMcmcInsertDelete.

NetworkPopulations.draw_sample — Method

draw_sample(
    mcmc::SisMcmcInsertDelete, 
    model::SIS;
    desired_samples::Int=mcmc.desired_samples, 
    burn_in::Int=mcmc.burn_in,
    lag::Int=mcmc.lag,
    init::Vector{Path{T}}=get_init(model, mcmc.init)
    )

NetworkPopulations.draw_sample — Method

draw_sample(
    mcmc::SisMcmcInsertDelete, 
    model::SIS;
    desired_samples::Int=mcmc.desired_samples, 
    burn_in::Int=mcmc.burn_in,
    lag::Int=mcmc.lag,
    init::Vector{Path{T}}=get_init(model, mcmc.init)
    )

NetworkPopulations.eachinterseq — Method

eachinterseq(V, K, L)

Returns an iterator over all interaction sequences over the vertex set V, with dimension bounds defined by K and L, specifically

V = vertex set, must be a vector of unique strings or integers;

K = max interaction length, must be an integer;

L = max number of interactions, must be an integer.

NetworkPopulations.get_normalising_const — Method

Calculate the normalising constant of SIS

NetworkPopulations.get_normalising_const — Method

Calculate the true normalising constant.

NetworkPopulations.get_path_sequences — Method

Given a DataFrames.GroupedDataFrame object and some time threshold of type Dates.TimePeriod this will convert the data to interaction sequences.

Column names for vertex labels and timestamps are passed names arguments.

Note the time column must be of type Dates.DateTime.

NetworkPopulations.get_sample_space — Method

Returns vector with all elements in the sample space.

NetworkPopulations.get_sample_space — Method

Returns vector with all elements in the sample space.

NetworkPopulations.get_vertex_counts — Method

Takes a GroupedDataFrame and a string indicating column name for vertex labels, outputs a Vector{Multiset}, that is, a multisets of vertex counts for each group in the data.

NetworkPopulations.multigraph_edge_list — Method

multigraph_edge_list(x::Path)

This is a helper function which can be used in conjunction with GraphRecipes.graphplot() to visualise a path.

Output is tuple (edge_list, label_map) where

edge_list::Vector{Vector{Int}} describes all edges, whereby edgelist[1] stores all edges starting at node 1, edgelist[2] those starting at 2 and so on...

label_map::Dict this maps the integer labels of nodes back to the of the input data, e.g. label_map[1] = "a" means node 1 represents the vertex "a" in the path p.

To plot path with labels call

graphplot(edge_list, names=label_map)

NetworkPopulations.pmf_unormalised — Method

Evaluate the unormalised probability of an interaction seq x

NetworkPopulations.string_to_int — Method

string_to_int(::InteractionSequence{String}, ::Dict{String,Int})

Takes `S::InteractionSequence{String}` and outputs a value of type `InteractionSequence{Int}` by applying the map encoded via the dictionary `mapper::Dict{String,Int}`.