NetworkPopulations.InitMode — Type

InitMode <: McmcInitialiser - this is a MCMC initialisation scheme for SIS model samplers which starts the MCMC chain at the model mode by default.

NetworkPopulations.McmcInitialiser — Type

Abstract type representing initialisation schemes for SIS model samplers.

NetworkPopulations.PosteriorPredictive — Type

A wrapper for the posterior predictive.

Note: the field model_type is used for constructing models.

NetworkPopulations.ScaledBeta — Type

Scaled Beta distribution. Instantiate with ScaledBeta() e.g.

d = ScaledBeta(α, β, 0.0, 10.0)

would make a ScaledBeta type d which is bounded by 0.0 and 10.0. Note

(α,β) are usual parameters;

lb = lower bound;

ub = upper bound.

NetworkPopulations.SingleMissingPredictive — Type

A predictive distribution for a single missing entry.

NetworkPopulations.TrGeometric — Type

TrGeomtric(p::Float64, lb::Int, ub::Int)

Truncated Geometric distribution. Note:

lb = lower bound = samllest possible value (default=0)

ub = upper bound = largest possible value.

Instantiation:

TrGeometric(p, ub) - here lb defaults to 0

TrGeometric(p, lb, ub) here lb is specified

NetworkPopulations.SimMcmcInsertDelete — Method

A constructor for a now depereciated sampler class, defined for backwards compatability. Will return the equivalent class in the current type hierarchy.

This regards a sampler for the SIM model combining edit allocation move with the path insertion deletion move.

NetworkPopulations.SisMcmcInsertDelete — Method

A constructor for a now depereciated sampler class, defined for backwards compatability. Will return the equivalent class in the current type hierarchy.

This regards a sampler for the SIS model combining edit allocation move with the path insertion deletion move.

NetworkPopulations.cardinality — Method

Calculate the sample space cardinality.

NetworkPopulations.cardinality — Method

Calculate the sample space cardinality.

NetworkPopulations.draw_sample — Method

draw_sample(mcmc::InvMcmcSampler, predictive::PosteriorPredictive; kwargs...)

Draw samples from the posterior predictive using the given iMCMC sampler. This involves two steps

Draw parameters from posterior (mode and γ)

Draw sample from model at these parameter values

This has also two keyword arguments

n_samples = number of draws from the posterior (mode and γ combinations)

n_reps = number of samples from model at each combination of mode and γ

NetworkPopulations.draw_sample_predictive — Method

draw_sample_predictive(mcmc::InvMcmcSampler, posterior_out::PosteriorMcmcOutput; kwargs...)

Draw samples from the posterior predictive using the given iMCMC sampler. This involves two steps

Draw parameters from posterior (mode and γ)

Draw sample from model at these parameter values

This has also two keyword arguments

n_samples = number of draws from the posterior (mode and γ combinations)

n_reps = number of samples from model at each combination of mode and γ

NetworkPopulations.eachinterseq — Method

eachinterseq(V, K, L)

Returns an iterator over all interaction sequences over the vertex set V, with dimension bounds defined by K and L, specifically

V = vertex set, must be a vector of unique strings or integers;

K = max interaction length, must be an integer;

L = max number of interactions, must be an integer.

NetworkPopulations.eval_posterior_predictive — Method

eval_posterior_predictive(S::InterSeq{Int}, ind::Tuple{Int,Int}, S_sample::InterSeqSample{Int}, γ_sample::Vector{Float64}, d::SemiMetric, V::UnitRange)

Evaluating posterior predictive of single missing entry. Takes in single observation S and index ind of missing entry, along with posterior samples for the mode S_sample and dispersion γ_sample, distance metric d and vertex set V, returning a instance of SingleMissingPredictive, which can be used to make predictions etc.

NetworkPopulations.get_normalising_const — Method

Calculate the true normalising constant.

NetworkPopulations.get_pred_accuracy — Method

get_pred_accuracy(preds::Vector{SingleMissingPredictive})

Given a vector of missing entry predictives, evaluate the predictive accuracy observed, that is, the proportion of times the correct prediction was made.

NetworkPopulations.get_prediction — Method

get_prediction(predictive::SingleMissingPredictive)

Query missing entry predictive for predicticted values.

Returns a vector of MAP values (can be more than one if there is a tie).

NetworkPopulations.get_prediction_rand — Method

get_prediction_rand(predictive::SingleMissingPredictive)

Query missing entry predictive for predicticted value. Returns tuple of (pred_val, entropy) containing the prediction and entropy of the predictive distribution.

Note: if there is a tie, we sample randomly.

NetworkPopulations.get_sample_space — Method

Returns vector with all elements in the sample space.

NetworkPopulations.get_sample_space — Method

Returns vector with all elements in the sample space.

NetworkPopulations.get_truth — Method

get_truth(pred::SingleMissingPredictive)

Query true entry for missing entry pred.

NetworkPopulations.multigraph_edge_list — Method

multigraph_edge_list(x::Path)

This is a helper function which can be used in conjunction with GraphRecipes.graphplot() to visualise a path.

Output is tuple (edge_list, label_map) where

edge_list::Vector{Vector{Int}} describes all edges, whereby edgelist[1] stores all edges starting at node 1, edgelist[2] those starting at 2 and so on...

label_map::Dict this maps the integer labels of nodes back to the of the input data, e.g. label_map[1] = "a" means node 1 represents the vertex "a" in the path p.

To plot path with labels call

graphplot(edge_list, names=label_map)

NetworkPopulations.pmf_unormalised — Method

Evaluate the unormalised probability of an interaction seq x

NetworkPopulations.pred_missing — Method

pred_missing(S::InterSeq{Int}, ind::Tuple{Int,Int}, posterior_out::PosteriorMcmcOutput)

Posterior predictive for missing entry. Returns instance of SingleMissingPredictive for predicting the entry of S indexed by ind.

NetworkPopulations.pred_missing — Method

pred_missing(S::InterSeq{Int}, ind::Tuple{Int,Int}, model::Union{SIS,SIM})

True model-based predictive for missing entry. Returns instance of SingleMissingPredictive for predicting the entry of S indexed by ind.

NetworkPopulations.string_to_int — Method

string_to_int(::InteractionSequence{String}, ::Dict{String,Int})

Takes `S::InteractionSequence{String}` and outputs a value of type `InteractionSequence{Int}` by applying the map encoded via the dictionary `mapper::Dict{String,Int}`.

NetworkPopulations.was_correct — Method

was_correct(pred::SingleMissingPredictive)

Test whether prediction was correct. Returns Boolean.